Prefer a static reducer (#4)
A `static` reducer will not only improve the code because it leaves less
room for side effects, it will also make individual actions easier to
test.

Side effects still should have access to the instance (e.g. to access
refs). To enable this, all side effects will now be called with a
reference to the component as the first argument.

Previous, class property reducers will still work until the 1.0.0
release to avoid making a breaking change in a minor version.